export/fuse: Fix termination of FUSE export under fuse-over-io_uring mode
When the fuse-over-io_uring mode is enabled, we submit a ring entry for
a commit-and-fetch operation to the kernel in the final step of the CQE
handler. The kernel may then wait indefinitely for new FUSE requests,
leaving the entries in a pending state. However, during this wait, the
FUSE export might need to be deleted. To prevent a situation where a CQE
returns after the FUSE export has been deleted and thus cannot be
processed, we use the block reference counter to ensure that the export
is not deleted while waiting for the CQE.